requirements.txt
flask>=3.0.0
flask-cors>=5.0.0
chromadb>=1.0.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
openai>=1.0.0
umap-learn>=0.5.0
hdbscan>=0.8.0
tenacity>=8.2.0
python-dotenv>=0.21.0
pandas>=2.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
tiktoken>=0.8.0
loguru>=0.7.0 
streamlit
faiss-cpu